Updated put package endpoint
